LayoutTests/imported/w3c/web-platform-tests/css/css-break/table/table-parts-offsets-vertical-rl.tentative-expected.txt


FAIL table assert_equals: offsetTop expected 8 but got 408
FAIL colgroup assert_equals: offsetTop expected 18 but got 411
FAIL col assert_equals: offsetTop expected 18 but got 411
FAIL col2 assert_equals: offsetTop expected 75 but got 411
FAIL colgroup2 assert_equals: offsetTop expected 132 but got 411
FAIL col3 assert_equals: offsetTop expected 132 but got 411
FAIL rowgroup assert_equals: offsetTop expected 18 but got 411
FAIL row assert_equals: offsetTop expected 18 but got 211
FAIL cell assert_equals: offsetTop expected 18 but got 218
FAIL content assert_equals: offsetTop expected 18 but got 218
FAIL cell2 assert_equals: offsetTop expected 75 but got 275
FAIL content2 assert_equals: offsetTop expected 75 but got 275
FAIL cell3 assert_equals: offsetTop expected 132 but got 332
FAIL content3 assert_equals: offsetTop expected 132 but got 332
FAIL row2 assert_equals: offsetTop expected 218 but got 411
FAIL cell4 assert_equals: offsetTop expected 218 but got 418
FAIL content4 assert_equals: offsetTop expected 218 but got 418
